feat: 기본 경로를 새 작업 화면으로 전환
