+++ src/app/layout.tsx
@@ -1,20 +1,7 @@
 import type { Metadata } from 'next'
 
-import dynamic from 'next/dynamic'
-
-import { Analytics } from '@vercel/analytics/next'
-
-import { ThemeProvider } from '@/shared/lib/theme-provider'
-
 import './globals.css'
-
-const Signature = dynamic(
-  () =>
-    import('@/shared/ui/animated-ui/signature').then((mod) => mod.Signature),
-  {
-    ssr: true,
-  }
-)
+import './theme.css'
 
 export const metadata: Metadata = {
   title: {
@@ -22,14 +9,12 @@ export const metadata: Metadata = {
     template: '%s | 트럭 매물 수집기',
   },
   description:
-    '중고 트럭 매물 정보와 이미지를 자동으로 수집하고 정리하는 웹 애플리케이션입니다. URL 입력만으로 트럭 매물 데이터를 자동 추출하여 체계적으로 관리할 수 있습니다.',
+    '중고 트럭 매물 주소를 빠르게 확인하고 차량별 이미지 폴더로 정리하는 작업 화면입니다.',
   keywords: [
     '트럭',
     '중고트럭',
     '매물수집',
     '데이터수집',
-    '자동화',
-    '웹스크래핑',
     '매물정보',
     '트럭매매',
     '상용차',
@@ -55,14 +40,14 @@ export const metadata: Metadata = {
     url: '/',
     title: '트럭 매물 수집기',
     description:
-      '중고 트럭 매물 정보와 이미지를 자동으로 수집하고 정리하는 웹 애플리케이션',
+      '중고 트럭 매물 주소를 확인하고 차량별 이미지 폴더로 정리합니다.',
     siteName: '트럭 매물 수집기',
     images: [
       {
         url: '/opengraph-image',
         width: 1200,
         height: 630,
-        alt: '트럭 매물 수집기 - 중고 트럭 매물 정보와 이미지를 자동 수집',
+        alt: '트럭 매물 수집기',
       },
     ],
   },
@@ -70,7 +55,7 @@ export const metadata: Metadata = {
     card: 'summary_large_image',
     title: '트럭 매물 수집기',
     description:
-      '중고 트럭 매물 정보와 이미지를 자동으로 수집하고 정리하는 웹 애플리케이션',
+      '중고 트럭 매물 주소를 확인하고 차량별 이미지 폴더로 정리합니다.',
     images: ['/opengraph-image'],
   },
   robots: {
@@ -84,10 +69,6 @@ export const metadata: Metadata = {
       'max-snippet': -1,
     },
   },
-  verification: {
-    // Google Search Console 인증 시 추가 (현재는 불필요)
-    // google: 'your-google-verification-code',
-  },
   category: 'productivity',
 }
 
@@ -97,18 +78,9 @@ export default function RootLayout({
   children: React.ReactNode
 }>) {
   return (
-    <html lang="ko" suppressHydrationWarning>
+    <html lang="ko">
       <body className="antialiased">
-        <ThemeProvider
-          attribute="class"
-          defaultTheme="system"
-          enableSystem
-          disableTransitionOnChange
-        >
-          <Signature />
-          {children}
-          <Analytics />
-        </ThemeProvider>
+        <div className="v2-theme min-h-dvh">{children}</div>
       </body>
     </html>
   )
